+++ flash/core/data/data_module.py
@@ -40,7 +40,7 @@ from flash.core.data.auto_dataset import BaseAutoDataset, IterableAutoDataset
 from flash.core.data.base_viz import BaseVisualization
 from flash.core.data.callback import BaseDataFetcher
 from flash.core.data.data_pipeline import DataPipeline
-from flash.core.data.data_source import DataSource, DefaultDataSources
+from flash.core.data.io.input import Input, InputFormat
 from flash.core.data.io.input_transform import DefaultInputTransform, InputTransform
 from flash.core.data.io.output_transform import OutputTransform
 from flash.core.data.splits import SplitDataset
@@ -56,7 +56,7 @@ else:
 
 class DataModule(pl.LightningDataModule):
     """A basic DataModule class for all Flash tasks. This class includes references to a
-    :class:`~flash.core.data.data_source.DataSource`, :class:`~flash.core.data.io.input_transform.InputTransform`,
+    :class:`~flash.core.data.io.input.Input`, :class:`~flash.core.data.io.input_transform.InputTransform`,
     :class:`~flash.core.data.io.output_transform.OutputTransform`, and a
     :class:`~flash.core.data.callback.BaseDataFetcher`.
 
@@ -65,7 +65,7 @@ class DataModule(pl.LightningDataModule):
         val_dataset: Dataset for validating model performance during training. Defaults to None.
         test_dataset: Dataset to test model performance. Defaults to None.
         predict_dataset: Dataset for predicting. Defaults to None.
-        data_source: The :class:`~flash.core.data.data_source.DataSource` that was used to create the datasets.
+        input: The :class:`~flash.core.data.io.input.Input` that was used to create the datasets.
         input_transform: The :class:`~flash.core.data.io.input_transform.InputTransform` to use when constructing the
             :class:`~flash.core.data.data_pipeline.DataPipeline`. If ``None``, a
             :class:`~flash.core.data.io.input_transform.DefaultInputTransform` will be used.
@@ -94,7 +94,7 @@ class DataModule(pl.LightningDataModule):
         val_dataset: Optional[Dataset] = None,
         test_dataset: Optional[Dataset] = None,
         predict_dataset: Optional[Dataset] = None,
-        data_source: Optional[DataSource] = None,
+        input: Optional[Input] = None,
         input_transform: Optional[InputTransform] = None,
         output_transform: Optional[OutputTransform] = None,
         data_fetcher: Optional[BaseDataFetcher] = None,
@@ -109,7 +109,7 @@ class DataModule(pl.LightningDataModule):
         if flash._IS_TESTING and torch.cuda.is_available():
             batch_size = 16
 
-        self._data_source: DataSource = data_source
+        self._input: Input = input
         self._input_tranform: Optional[InputTransform] = input_transform
         self._output_transform: Optional[OutputTransform] = output_transform
         self._viz: Optional[BaseVisualization] = None
@@ -426,9 +426,9 @@ class DataModule(pl.LightningDataModule):
         return multi_label_train or multi_label_val or multi_label_test
 
     @property
-    def data_source(self) -> Optional[DataSource]:
+    def input(self) -> Optional[Input]:
         """Property that returns the data source."""
-        return self._data_source
+        return self._input
 
     @property
     def input_transform(self) -> InputTransform:
@@ -445,16 +445,16 @@ class DataModule(pl.LightningDataModule):
     def data_pipeline(self) -> DataPipeline:
         """Property that returns the full data pipeline including the data source, input transform and
         postprocessing."""
-        return DataPipeline(self.data_source, self.input_transform, self.output_transform)
+        return DataPipeline(self.input, self.input_transform, self.output_transform)
 
-    def available_data_sources(self) -> Sequence[str]:
+    def available_inputs(self) -> Sequence[str]:
         """Get the list of available data source names for use with this
         :class:`~flash.core.data.data_module.DataModule`.
 
         Returns:
             The list of data source names.
         """
-        return self.input_transform.available_data_sources()
+        return self.input_transform.available_inputs()
 
     @staticmethod
     def _split_train_val(
@@ -492,9 +492,9 @@ class DataModule(pl.LightningDataModule):
         )
 
     @classmethod
-    def from_data_source(
+    def from_input(
         cls,
-        data_source: str,
+        input: str,
         train_data: Any = None,
         val_data: Any = None,
         test_data: Any = None,
@@ -512,21 +512,21 @@ class DataModule(pl.LightningDataModule):
         **input_transform_kwargs: Any,
     ) -> "DataModule":
         """Creates a :class:`~flash.core.data.data_module.DataModule` object from the given inputs to
-        :meth:`~flash.core.data.data_source.DataSource.load_data` (``train_data``, ``val_data``, ``test_data``,
+        :meth:`~flash.core.data.io.input.Input.load_data` (``train_data``, ``val_data``, ``test_data``,
         ``predict_data``). The data source will be resolved from the instantiated
         :class:`~flash.core.data.io.input_transform.InputTransform`
-        using :meth:`~flash.core.data.io.input_transform.InputTransform.data_source_of_name`.
+        using :meth:`~flash.core.data.io.input_transform.InputTransform.input_of_name`.
 
         Args:
-            data_source: The name of the data source to use for the
-                :meth:`~flash.core.data.data_source.DataSource.load_data`.
-            train_data: The input to :meth:`~flash.core.data.data_source.DataSource.load_data` to use when creating
+            input: The name of the data source to use for the
+                :meth:`~flash.core.data.io.input.Input.load_data`.
+            train_data: The input to :meth:`~flash.core.data.io.input.Input.load_data` to use when creating
                 the train dataset.
-            val_data: The input to :meth:`~flash.core.data.data_source.DataSource.load_data` to use when creating
+            val_data: The input to :meth:`~flash.core.data.io.input.Input.load_data` to use when creating
                 the validation dataset.
-            test_data: The input to :meth:`~flash.core.data.data_source.DataSource.load_data` to use when creating
+            test_data: The input to :meth:`~flash.core.data.io.input.Input.load_data` to use when creating
                 the test dataset.
-            predict_data: The input to :meth:`~flash.core.data.data_source.DataSource.load_data` to use when creating
+            predict_data: The input to :meth:`~flash.core.data.io.input.Input.load_data` to use when creating
                 the predict dataset.
             train_transform: The dictionary of transforms to use during training which maps
                 :class:`~flash.core.data.io.input_transform.InputTransform` hook names to callable transforms.
@@ -553,8 +553,8 @@ class DataModule(pl.LightningDataModule):
 
         Examples::
 
-            data_module = DataModule.from_data_source(
-                DefaultDataSources.FOLDERS,
+            data_module = DataModule.from_input(
+                InputFormat.FOLDERS,
                 train_data="train_folder",
                 train_transform={
                     "to_tensor_transform": torch.as_tensor,
@@ -570,9 +570,9 @@ class DataModule(pl.LightningDataModule):
             **input_transform_kwargs,
         )
 
-        data_source = input_transform.data_source_of_name(data_source)
+        input = input_transform.input_of_name(input)
 
-        train_dataset, val_dataset, test_dataset, predict_dataset = data_source.to_datasets(
+        train_dataset, val_dataset, test_dataset, predict_dataset = input.to_datasets(
             train_data,
             val_data,
             test_data,
@@ -584,7 +584,7 @@ class DataModule(pl.LightningDataModule):
             val_dataset,
             test_dataset,
             predict_dataset,
-            data_source=data_source,
+            input=input,
             input_transform=input_transform,
             data_fetcher=data_fetcher,
             val_split=val_split,
@@ -613,8 +613,8 @@ class DataModule(pl.LightningDataModule):
         **input_transform_kwargs: Any,
     ) -> "DataModule":
         """Creates a :class:`~flash.core.data.data_module.DataModule` object from the given folders using the
-        :class:`~flash.core.data.data_source.DataSource` of name
-        :attr:`~flash.core.data.data_source.DefaultDataSources.FOLDERS`
+        :class:`~flash.core.data.io.input.Input` of name
+        :attr:`~flash.core.data.io.input.InputFormat.FOLDERS`
         from the passed or constructed :class:`~flash.core.data.io.input_transform.InputTransform`.
 
         Args:
@@ -645,8 +645,8 @@ class DataModule(pl.LightningDataModule):
         Returns:
             The constructed data module.
         """
-        return cls.from_data_source(
-            DefaultDataSources.FOLDERS,
+        return cls.from_input(
+            InputFormat.FOLDERS,
             train_folder,
             val_folder,
             test_folder,
@@ -687,8 +687,8 @@ class DataModule(pl.LightningDataModule):
         **input_transform_kwargs: Any,
     ) -> "DataModule":
         """Creates a :class:`~flash.core.data.data_module.DataModule` object from the given sequences of files
-        using the :class:`~flash.core.data.data_source.DataSource` of name
-        :attr:`~flash.core.data.data_source.DefaultDataSources.FILES` from the passed or constructed
+        using the :class:`~flash.core.data.io.input.Input` of name
+        :attr:`~flash.core.data.io.input.InputFormat.FILES` from the passed or constructed
         :class:`~flash.core.data.io.input_transform.InputTransform`.
 
         Args:
@@ -722,8 +722,8 @@ class DataModule(pl.LightningDataModule):
         Returns:
             The constructed data module.
         """
-        return cls.from_data_source(
-            DefaultDataSources.FILES,
+        return cls.from_input(
+            InputFormat.FILES,
             (train_files, train_targets),
             (val_files, val_targets),
             (test_files, test_targets),
@@ -764,8 +764,8 @@ class DataModule(pl.LightningDataModule):
         **input_transform_kwargs: Any,
     ) -> "DataModule":
         """Creates a :class:`~flash.core.data.data_module.DataModule` object from the given tensors using the
-        :class:`~flash.core.data.data_source.DataSource`
-        of name :attr:`~flash.core.data.data_source.DefaultDataSources.TENSOR`
+        :class:`~flash.core.data.io.input.Input`
+        of name :attr:`~flash.core.data.io.input.InputFormat.TENSOR`
         from the passed or constructed :class:`~flash.core.data.io.input_transform.InputTransform`.
 
         Args:
@@ -809,8 +809,8 @@ class DataModule(pl.LightningDataModule):
                 },
             )
         """
-        return cls.from_data_source(
-            DefaultDataSources.TENSORS,
+        return cls.from_input(
+            InputFormat.TENSORS,
             (train_data, train_targets),
             (val_data, val_targets),
             (test_data, test_targets),
@@ -851,8 +851,8 @@ class DataModule(pl.LightningDataModule):
         **input_transform_kwargs: Any,
     ) -> "DataModule":
         """Creates a :class:`~flash.core.data.data_module.DataModule` object from the given numpy array using the
-        :class:`~flash.core.data.data_source.DataSource`
-        of name :attr:`~flash.core.data.data_source.DefaultDataSources.NUMPY`
+        :class:`~flash.core.data.io.input.Input`
+        of name :attr:`~flash.core.data.io.input.InputFormat.NUMPY`
         from the passed or constructed :class:`~flash.core.data.io.input_transform.InputTransform`.
 
         Args:
@@ -896,8 +896,8 @@ class DataModule(pl.LightningDataModule):
                 },
             )
         """
-        return cls.from_data_source(
-            DefaultDataSources.NUMPY,
+        return cls.from_input(
+            InputFormat.NUMPY,
             (train_data, train_targets),
             (val_data, val_targets),
             (test_data, test_targets),
@@ -938,8 +938,8 @@ class DataModule(pl.LightningDataModule):
         **input_transform_kwargs: Any,
     ) -> "DataModule":
         """Creates a :class:`~flash.core.data.data_module.DataModule` object from the given JSON files using the
-        :class:`~flash.core.data.data_source.DataSource`
-        of name :attr:`~flash.core.data.data_source.DefaultDataSources.JSON`
+        :class:`~flash.core.data.io.input.Input`
+        of name :attr:`~flash.core.data.io.input.InputFormat.JSON`
         from the passed or constructed :class:`~flash.core.data.io.input_transform.InputTransform`.
 
         Args:
@@ -1006,8 +1006,8 @@ class DataModule(pl.LightningDataModule):
                 feild="data"
             )
         """
-        return cls.from_data_source(
-            DefaultDataSources.JSON,
+        return cls.from_input(
+            InputFormat.JSON,
             (train_file, input_fields, target_fields, field),
             (val_file, input_fields, target_fields, field),
             (test_file, input_fields, target_fields, field),
@@ -1047,8 +1047,8 @@ class DataModule(pl.LightningDataModule):
         **input_transform_kwargs: Any,
     ) -> "DataModule":
         """Creates a :class:`~flash.core.data.data_module.DataModule` object from the given CSV files using the
-        :class:`~flash.core.data.data_source.DataSource`
-        of name :attr:`~flash.core.data.data_source.DefaultDataSources.CSV`
+        :class:`~flash.core.data.io.input.Input`
+        of name :attr:`~flash.core.data.io.input.InputFormat.CSV`
         from the passed or constructed :class:`~flash.core.data.io.input_transform.InputTransform`.
 
         Args:
@@ -1092,8 +1092,8 @@ class DataModule(pl.LightningDataModule):
                 },
             )
         """
-        return cls.from_data_source(
-            DefaultDataSources.CSV,
+        return cls.from_input(
+            InputFormat.CSV,
             (train_file, input_fields, target_fields),
             (val_file, input_fields, target_fields),
             (test_file, input_fields, target_fields),
@@ -1131,8 +1131,8 @@ class DataModule(pl.LightningDataModule):
         **input_transform_kwargs: Any,
     ) -> "DataModule":
         """Creates a :class:`~flash.core.data.data_module.DataModule` object from the given datasets using the
-        :class:`~flash.core.data.data_source.DataSource`
-        of name :attr:`~flash.core.data.data_source.DefaultDataSources.DATASETS`
+        :class:`~flash.core.data.io.input.Input`
+        of name :attr:`~flash.core.data.io.input.InputFormat.DATASETS`
         from the passed or constructed :class:`~flash.core.data.io.input_transform.InputTransform`.
 
         Args:
@@ -1172,8 +1172,8 @@ class DataModule(pl.LightningDataModule):
                 },
             )
         """
-        return cls.from_data_source(
-            DefaultDataSources.DATASETS,
+        return cls.from_input(
+            InputFormat.DATASETS,
             train_dataset,
             val_dataset,
             test_dataset,
@@ -1212,8 +1212,8 @@ class DataModule(pl.LightningDataModule):
     ) -> "DataModule":
         """Creates a :class:`~flash.core.data.data_module.DataModule` object
         from the given FiftyOne Datasets using the
-        :class:`~flash.core.data.data_source.DataSource` of name
-        :attr:`~flash.core.data.data_source.DefaultDataSources.FIFTYONE`
+        :class:`~flash.core.data.io.input.Input` of name
+        :attr:`~flash.core.data.io.input.InputFormat.FIFTYONE`
         from the passed or constructed :class:`~flash.core.data.io.input_transform.InputTransform`.
 
         Args:
@@ -1256,8 +1256,8 @@ class DataModule(pl.LightningDataModule):
                 },
             )
         """
-        return cls.from_data_source(
-            DefaultDataSources.FIFTYONE,
+        return cls.from_input(
+            InputFormat.FIFTYONE,
             train_dataset,
             val_dataset,
             test_dataset,
@@ -1300,8 +1300,8 @@ class DataModule(pl.LightningDataModule):
     ) -> "DataModule":
         """Creates a :class:`~flash.core.data.data_module.DataModule` object
         from the given export file and data directory using the
-        :class:`~flash.core.data.data_source.DataSource` of name
-        :attr:`~flash.core.data.data_source.DefaultDataSources.FOLDERS`
+        :class:`~flash.core.data.io.input.Input` of name
+        :attr:`~flash.core.data.io.input.InputFormat.FOLDERS`
         from the passed or constructed :class:`~flash.core.data.io.input_transform.InputTransform`.
 
         Args:
@@ -1381,8 +1381,8 @@ class DataModule(pl.LightningDataModule):
                 "export_json": predict_export_json,
                 "multi_label": input_transform_kwargs.get("multi_label", False),
             }
-        return cls.from_data_source(
-            DefaultDataSources.LABELSTUDIO,
+        return cls.from_input(
+            InputFormat.LABELSTUDIO,
             train_data=train_data if train_data else data,
             val_data=val_data,
             test_data=test_data,