+++ flash/image/face_detection/model.py
@@ -16,7 +16,7 @@ from typing import Any, Dict, Iterable, List, Union
 import torch
 from torch.nn import Module
 
-from flash.core.data.data_source import DefaultDataKeys
+from flash.core.data.io.input import DataKeys
 from flash.core.data.io.output import Output
 from flash.core.model import Task
 from flash.core.utilities.imports import _FASTFACE_AVAILABLE
@@ -39,7 +39,7 @@ class DetectionLabels(Output):
     """A :class:`.Output` which extracts predictions from sample dict."""
 
     def transform(self, sample: Any) -> Dict[str, Any]:
-        return sample[DefaultDataKeys.PREDS] if isinstance(sample, Dict) else sample
+        return sample[DataKeys.PREDS] if isinstance(sample, Dict) else sample
 
 
 class FaceDetector(Task):
@@ -145,7 +145,7 @@ class FaceDetector(Task):
             metric.update(pred_boxes, target_boxes)
 
     def __shared_step(self, batch, train=False) -> Any:
-        images, targets = batch[DefaultDataKeys.INPUT], batch[DefaultDataKeys.TARGET]
+        images, targets = batch[DataKeys.INPUT], batch[DataKeys.TARGET]
         images = self._prepare_batch(images)
         logits = self.model(images)
         loss = self.model.compute_loss(logits, targets)
@@ -181,8 +181,8 @@ class FaceDetector(Task):
         self.log_dict({f"test_{k}": v for k, v in metric_results.items()}, on_epoch=True)
 
     def predict_step(self, batch: Any, batch_idx: int, dataloader_idx: int = 0) -> Any:
-        images = batch[DefaultDataKeys.INPUT]
-        batch[DefaultDataKeys.PREDS] = self(images)
+        images = batch[DataKeys.INPUT]
+        batch[DataKeys.PREDS] = self(images)
         return batch
 
     def modules_to_freeze(self) -> Union[Module, Iterable[Union[Module, Iterable]]]: